Combine root and dbname in app.js
This first step hoists the step of generating the database URL into
app.js rather than each of the backup and restore files. The functions
called by app.js all operate on a single database, so need not have
separate root and dbname supplied.

#26
